+++ src/mfx/dsp/dly/DelayLineReader.h
@@ -52,19 +52,19 @@ public:
 	               DelayLineReader ()  = default;
 	virtual        ~DelayLineReader () = default;
 
-	void           set_tmp_buf (float *buf_ptr, long len);
+	void           set_tmp_buf (float *buf_ptr, int len);
 	float *        get_tmp_buf_ptr () const;
-	long           get_tmp_buf_len () const;
+	int            get_tmp_buf_len () const;
 
 	void           set_delay_line (DelayLine &delay_line);
 
 	void           set_resampling_range (double rate_inf, double rate_sup);
-	void           set_crossfade (long nbr_spl, const float shape_ptr []);
+	void           set_crossfade (int nbr_spl, const float shape_ptr []);
 
 	bool           is_ready () const;
 
-	void           set_delay_time (double delay_time, long transition_time);
-	void           read_data (float dest_ptr [], long nbr_spl, long src_pos);
+	void           set_delay_time (double delay_time, int transition_time);
+	void           read_data (float dest_ptr [], int nbr_spl, int src_pos);
 	bool           is_time_ramping () const;
 	bool           is_time_change_programmed () const;
 	void           clear_buffers ();
@@ -81,15 +81,15 @@ protected:
 
 private:
 
-	void           setup_immediate_transition (double delay_time, long transition_time);
-	void           apply_crossfade (float dest_ptr [], long nbr_spl, double lerp_pos_end, long src_pos);
+	void           setup_immediate_transition (double delay_time, int transition_time);
+	void           apply_crossfade (float dest_ptr [], int nbr_spl, double lerp_pos_end, int src_pos);
 
 	DelayLine *    _delay_line_ptr  =  0;  // 0 = not initialised.
 
 	const float *  _xfade_shape_ptr =  0; // 0 = not set.
 
 	float *        _tmp_buf_ptr     =  0;  // 0: not set.
-	long           _tmp_buf_len     =  0;  // samples, > 0
+	int            _tmp_buf_len     =  0;  // samples, > 0
 
 	double         _rate_inf   = -2;    // Out of the rate range, we crossfade instead
 	double         _rate_sup   = +2;    // of resampling data.
@@ -97,10 +97,10 @@ private:
 	double         _time_beg   =  0;    // Delay time of the ramp beginning, or delay time for the fading out part. s.
 	double         _time_end   =  0;    // Target delay time, or delay time for the fading in part. s.
 	double         _time_prog  =  0;    // Programmed delay time, if the user set a new delay time during a transition. s.
-	long           _trans_dur  =  0;    // Transition duration between two delay values (samples).
-	long           _trans_pos  = -1;    // Position within the transition (samples). Negative = no transition.
-	long           _trans_prog = -1;    // Number of samples before the programmed new delay time (sample). Negative = no programmed time change.
-	long           _xfade_dur  = 64;    // Crossfade duration, samples
+	int            _trans_dur  =  0;    // Transition duration between two delay values (samples).
+	int            _trans_pos  = -1;    // Position within the transition (samples). Negative = no transition.
+	int            _trans_prog = -1;    // Number of samples before the programmed new delay time (sample). Negative = no programmed time change.
+	int            _xfade_dur  = 64;    // Crossfade duration, samples
 	bool           _xfade_flag = false; // Current transition is a crossfade.
